A lot of documentation has been added
Also, the cookie store / jar makes a bit more sense. It was very
difficult to create a store and jar, but now, we can simply add the
store to the zone_builder and it works. If you are going to add a
specific jar, you best get them from a store, otherwise you have to
convert it into an Handle (arc/rwlock), which is not hard, but a bit
cumbersome to do..  It's fine for now