Implemented a vello renderer
The renderer connects with the wgpu from the frontend (eframe) through a
wgpu context provider.

We also use a simple fontique/parley system for drawing fonts. There is
a lot of optimization that needs to be done here.

As far as we can see now, there is zero-copy of texture data. So all is
generated through the GPU and is painted through eframe without any
copies from GPU to CPU memory.