Create a resource fetch system
Instead of doing a very simple "reqwest" to fetch a resource, this
system will deal with the complexity of fetching resources. Multiple
tabs requesting the same resource, dealing with streams and buffered
requests, dealing with timeouts, cancellations by the user etc.

This all resulted in a rather complex setup that needs a bit of
documentation (hopefully will fix that fast enough).

This setup can function as a base system for the engine. I think it
should work pretty well, but more tests and benchmarks are needed.